workers/pdf-refinery/requirements-embed.txt
boto3
psycopg2-binary
azure-storage-blob
